Updated delay param to float in addAnimationForTrack

+++ spine-cocos2d-iphone/3/src/spine/SkeletonAnimation.m
@@ -178,7 +178,7 @@ void disposeTrackEntry (spTrackEntry* entry) {
 	return spAnimationState_setAnimation(_state, trackIndex, animation, loop);
 }
 
-- (spTrackEntry*) addAnimationForTrack:(int)trackIndex name:(NSString*)name loop:(bool)loop afterDelay:(int)delay {
+- (spTrackEntry*) addAnimationForTrack:(int)trackIndex name:(NSString*)name loop:(bool)loop afterDelay:(float)delay {
 	spAnimation* animation = spSkeletonData_findAnimation(_skeleton->data, [name UTF8String]);
 	if (!animation) {
 		CCLOG(@"Spine: Animation not found: %@", name);
